File structure refactor
