+++ src/shapes/mod.rs
@@ -1,9 +1,3 @@
 pub mod box_quad;
-pub mod hit;
-pub mod material;
 pub mod quad;
 pub mod sphere;
-pub mod transformation;
-
-pub use hit::*;
-pub use material::*;
