+++ src/shapes/quad.rs
@@ -1,7 +1,8 @@
 #![allow(unused)]
 
+use crate::hit::{HitRecord, Hittable};
+use crate::material::Material;
 use crate::ray::Ray;
-use crate::shapes::{HitRecord, Hittable, Material};
 use crate::vector3::Vector3;
 use std::sync::Arc;
