+++ src/transformation.rs
@@ -1,5 +1,5 @@
+use crate::hit::{HitRecord, Hittable};
 use crate::ray::Ray;
-use crate::shapes::{HitRecord, Hittable};
 use crate::vector3::Vector3;
 use std::sync::Arc;
